refactor(has_cache): extract cache to support
