+++ lib/convenient_service/common/plugins/has_internals/entities/internals/plugins/has_cache.rb
@@ -1,4 +1,3 @@
 # frozen_string_literal: true
 
 require_relative "has_cache/concern"
-require_relative "has_cache/entities"
