--- lib/convenient_service/common/plugins/has_internals/entities/internals/plugins/has_cache/entities.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "entities/cache"
