+++ lib/convenient_service/support.rb
@@ -3,6 +3,7 @@
 require_relative "support/concern"
 
 require_relative "support/abstract_method"
+require_relative "support/cache"
 require_relative "support/castable"
 require_relative "support/command"
 require_relative "support/copyable"
