+++ spec/lib/convenient_service/support/cache_spec.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require "spec_helper"
+
+require "convenient_service"
+
+##
+# TODO: Specs.
+#
+# RSpec.describe ConvenientService::Support::Cache do
+#   # ...
+# end
